Match sub_819AB78
Factory pipeline (tools/factory), source=tier2. Verified with a from-scratch build before commit, per the project's standing rule.

+++ src/sub_8047B5C.c
@@ -12,9 +12,9 @@ asm_unified(".include \"asm/nonmatching/sub_8047B5C.s\"");
 #else
 void sub_8047B5C(void *arg0, u16 arg1)
 {
-  unsigned long new_var;
-  new_var = *((u16 *) (((s8 *) arg0) + 0x266));
-  *((u16 *) (((s8 *) arg0) + 0x268)) = (u16) new_var;
+  u16 new_var;
+  new_var = (u16) (*((u16 *) (((s8 *) arg0) + 0x266)));
+  *((u16 *) (((s8 *) arg0) + 0x268)) = new_var;
   *((u16 *) (((s8 *) arg0) + 0x266)) = arg1;
 }
 #endif
